renamed package server to event
