+++ src/worker/event/lambdaServer.go
@@ -1,4 +1,4 @@
-package server
+package event
 
 import (
 	"fmt"
